feat: stream component with suspense
