+++ app/dashboard/(overview)/no_loading_.tsx
@@ -1,0 +1,5 @@
+import DashboardSkeleton from '../../ui/skeletons';
+
+export default function Loading() {
+  return <DashboardSkeleton />;
+}
