+++ app/dashboard/(overview)/page.tsx
@@ -1,0 +1,58 @@
+import { Suspense } from 'react';
+import CardWrapper from '@/app/ui/dashboard/cards';
+import RevenueChart from '@/app/ui/dashboard/revenue-chart';
+import LatestInvoices from '@/app/ui/dashboard/latest-invoices';
+import { lusitana } from '@/app/ui/fonts';
+import {
+  // fetchRevenue,
+  // fetchLatestInvoices,
+  fetchCardData,
+} from '@/app/lib/data';
+import {
+  RevenueChartSkeleton,
+  LatestInvoicesSkeleton,
+  CardSkeleton,
+} from '@/app/ui/skeletons';
+
+export default async function Page() {
+  // const revenue = await fetchRevenue(); // THis is blocking all other component thus the whole Page
+  // moving the fetching logic in a Suspense below will let the other componet render while RevenueChart is Loading
+  // for this to work we need to remove the loading.tsx component or rename it.
+
+  // const latestInvoices = await fetchLatestInvoices();
+  // const {
+  //   numberOfCustomers,
+  //   numberOfInvoices,
+  //   totalPaidInvoices,
+  //   totalPendingInvoices,
+  // } = await fetchCardData();
+
+  return (
+    <main>
+      <h1 className={`${lusitana.className} mb-4 text-xl md:text-2xl`}>
+        Dashboard
+      </h1>
+      <div className="grid gap-6 sm:grid-cols-2 lg:grid-cols-4">
+        <Suspense fallback={<CardSkeleton />}>
+          <CardWrapper />
+        </Suspense>
+        {/* <Card title="Collected" value={totalPaidInvoices} type="collected" />
+        <Card title="Pending" value={totalPendingInvoices} type="pending" />
+        <Card title="Total Invoices" value={numberOfInvoices} type="invoices" />
+        <Card
+          title="Total Customers"
+          value={numberOfCustomers}
+          type="customers"
+        /> */}
+      </div>
+      <div className="mt-6 grid grid-cols-1 gap-6 md:grid-cols-4 lg:grid-cols-8">
+        <Suspense fallback={<RevenueChartSkeleton />}>
+          <RevenueChart />
+        </Suspense>
+        <Suspense fallback={<LatestInvoicesSkeleton />}>
+          <LatestInvoices />
+        </Suspense>
+      </div>
+    </main>
+  );
+}
